// PruebaCargue/src/app/models/Res4505.ts
export interface Res4505 {
	PKId?: number;
    Tipo_Registro: string;
    Consecutivo_registro: string;
    Coidigohabilitacionips_primaria: string;
    Tipoid_usuario: string;
    Numero_Identificacion: string;
    PrimerApellido_usuario: string;
    SegundoApellido_usuario: string;
    PrimerNombre_usuario: string;
    SegundoNombre_usuario : string;
    Fecha_nacimiento:string;
    Sexo:string;
    Codigo_pertenenciaetnica:number;
    Codigo_ocupacion: number;
    Codigo_niveleducativo: number;
    Gestacion: number;
    Sifilisgestacional_congenita: number;
	Hipertension_Inducida_por_la_Gestacion: number;
    Hipotiroidismo_Congenito: number;
    Sintomatico_Respiratorio: number;
    Tuberculosis_Multidrogoresistente: number;
    Lepra: number;
    Obesidad_Desnutricion_Proteico_Calorica: number;
    Victima_Maltrato: number;
    Victima_Violencia_Sexual: number;
    Infecciones_TrasmisionSexual: number;
    Enfermedad_Mental: number;
    Cancer_Cervix: number;
    Cancer_Seno: number;
    Fluorosis_Dental: number;
    Fecha_Peso: string;
    Peso_Kilogramos: number;
    Fecha_Talla: string;
    Talla_Centimetros: number;
    Fecha_ProbableParto: string;
    Edad_GestacionalNacer: number;
	BCG: number;
    Hepatitis_B_menores1ano: number;
	Pentavalente: number;
	Polio: number;
	DPT_menores5anos: number;
	Rotavirus: number;
	Neumococo :number ;
	Influenza_Ninos: number ;
	Fiebre_Amarillaninosde1ano: number;
	HepatitisA: number;
	Triple_ViralNinos: number; 
	Virus_Papiloma_Humano: number; 
    TDTTMujeresEdad_Fertil_15_49_anos: number; 
	ControlPlaca_Bacteriana: number; 
	Fecha_atencion_parto_o_cesarea: string; 
	Fecha_salida_atencion_parto_cesarea: string ; 
	Fecha_consejeria_Lactancia_Materna: string; 
    Control_Recien_Nacido: string; 
	Planificacion_Familiar_Primeravez: string; 
	Suministro_Metodo_Anticonceptivo: number; 
	Fecha_Suministro_Metodo_Anticonceptivo: string; 
    ControlPrenatal_Primeravez: string; 
	Control_Prenatal: number; 
	ultimo_Control_Prenatal: string; 
	Suministroacido_Folico_ultimo_Control_Prenatal: number; 
	SuministroSulfato_Ferroso_ultimoControlPrenatal: number; 
	Suministro_CarbonatoCalcio_ltimoControlPrenatal: number; 
	Valoracion_Agudeza_Visual: string; 
	Consulta_por_Oftalmologia: string; 
	Fecha_Diagnostico_Desnutricion_Proteico_Calorica: string; 
	Consulta_Mujer_o_Menor_Victima_Maltrato: string; 
	Consulta_Victimas_Violencia_Sexual: string; 
	Consulta_Nutricion: string; 
	Consulta_de_Psicologia: string; 
	Consulta_Crecimiento_y_DesarrolloPrimeravez: string; 
	Suministro_Sulfato_Ferroso_ultima_Consulta_del_Menor_10_anos: number;
	Suministro_Vitamina_A_ultima_Consulta_Menor_10_anos: number; 
	Consulta_JovenPrimeravez: string; 
	Consulta_AdultoPrimeravez: string; 
	Preservativos_entregados_a_pacientes_con_ITS: number; 
	Asesoria_Pre_test_Elisa_para_VIH: string; 
	Asesoria_Pos_test_Elisa_para_VIH: string; 
	PCAEDCBAIC: number; 
	Fecha_Antigeno_Superficie_Hepatitis_B_en_Gestantes: string; 
	Resultado_Antigeno_Superficie_Hepatitis_B_en_Gestantes: number; 
    Fecha_Serologia_para_Sifilis: string; 
	Resultado_Serologia_para_Sifilis: number; 
	Fecha_Toma_Elisa_para_VIH: string; 
	Resultado_Elisa_para_VIH: number; 
	Fecha_TSH_Neonatal: string; 
	Resultado_TSH_Neonatal: number; 
	Tamizaje_Cancer_Cuello_Uterino: number; 
	Citologia_Cervico_uterina: string; 
	Citologia_Cervico_uterina_Resultados_segun_Bethesda: number; 
	Calidad_Muestra_Citologia_Cervicouterina: number; 
	Codigo_habilitacion_IPS_donde_se_toma_Citologia_Cervicouterina: string; 
	Fecha_Colposcopia: string; 
	Codigo_habilitacion_IPS_donde_se_toma_Colposcopia: string; 
	Fecha_biopsia_cervical: string; 
	Resultado_Biopsia_Cervical: number; 
	Codigo_habilitacion_IPS_donde_se_toma_biopsia: string; 
	Fecha_Mamografia: string; 
	Resultado_Mamografia: number; 
	Codigo_habilitacion_IPS_donde_se_toma_Mamografia: string; 
	Fecha_Toma_Biopsia_Seno_por_BACAF: string; 
	Fecha_Resultado_Biopsia_Seno_por_BACAF: string; 
	Biopsia_Seno_por_BACAF: string; 
	Codigo_habilitacion_IPS_donde_se_toma_Biopsia_Seno_por_BACAF: string; 
	Fecha_Toma_Hemoglobina: string; 
	Hemoglobina: string; 
	Fecha_Toma_Glicemia_Basalstring: string; 
	Fecha_Creatinina: string; 
	Creatinina: number; 
	Fecha_Hemoglobina_Glicosilada: string; 
	Hemoglobina_Glicosilada: number; 
	Fecha_Toma_Microalbuminuria: string; 
	Fecha_Toma_HDL: string; 
	Fecha_Toma_Baciloscopia_Diagnostico: string; 
	Baciloscopia_Diagnostico: number; 
	Tratamiento_Hipotiroidismo_Congenito: number; 
	Tratamiento_Sifilis_gestacional: number; 
	Tratamiento_Sifilis_Congenita:  number; 
	Tratamiento_Lepra:  number; 
	Fecha_Terminacion_Tratamiento_para_Leishmaniasis:string; 
}